[AbstractRector] add getClassName() method
