+++ src/Builder/Kernel/ServiceFromKernelResolver.php
@@ -2,23 +2,29 @@
 
 namespace Rector\Builder\Kernel;
 
+use PhpParser\Node\Arg;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
 final class ServiceFromKernelResolver
 {
-    public function resolveServiceClassByNameFromKernel(string $serviceName, string $kernelClass): ?string
+    public function resolveServiceClassFromArgument(Arg $argNode, string $kernelClass): ?string
     {
-        /** @var Kernel $kernel */
-        $kernel = new $kernelClass('dev', true);
-        $kernel->boot();
+        $serviceName = $argNode->value->value;
+        $serviceType = $this->resolveServiceClassByNameFromKernel($serviceName, $kernelClass);
 
-        // @todo: cache
-        // @todo: initialize without creating cache or log directory
-        // @todo: call only loadBundles() and initializeContainer() methods
+        if ($serviceType === null) {
+            return null;
+        }
 
-        /** @var ContainerInterface $container */
-        $container = $kernel->getContainer();
+        return $serviceType;
+    }
+
+
+    private function resolveServiceClassByNameFromKernel(string $serviceName, string $kernelClass): ?string
+    {
+        $container = $this->createContainerFromKernelClass($kernelClass);
+
         if (! $container->has($serviceName)) {
             // service name could not be found
             return null;
@@ -27,5 +33,25 @@ final class ServiceFromKernelResolver
         $service = $container->get($serviceName);
 
         return get_class($service);
+    }
+
+    private function createContainerFromKernelClass(string $kernelClass): ContainerInterface
+    {
+        $kernel = $this->createKernelFromKernelClass($kernelClass);
+
+        return $kernel->getContainer();
+    }
+
+    private function createKernelFromKernelClass(string $kernelClass): Kernel
+    {
+        /** @var Kernel $kernel */
+        $kernel = new $kernelClass('dev', true);
+        $kernel->boot();
+
+        return $kernel;
+
+        // @todo: cache
+        // @todo: initialize without creating cache or log directory
+        // @todo: call only loadBundles() and initializeContainer() methods
     }
 }
