+++ src/Rector/AbstractRector.php
@@ -3,6 +3,7 @@
 namespace Rector\Rector;
 
 use PhpParser\Node;
+use PhpParser\Node\Stmt\Class_;
 use PhpParser\NodeTraverser;
 use PhpParser\NodeVisitorAbstract;
 use Rector\Contract\Deprecation\DeprecationInterface;
@@ -10,6 +11,34 @@ use Rector\Contract\Rector\RectorInterface;
 
 abstract class AbstractRector extends NodeVisitorAbstract implements DeprecationInterface, RectorInterface
 {
+    /**
+     * @var Class_|null
+     */
+    protected $classNode;
+
+    /**
+     * @param Node[] $nodes
+     * @return null|Node[]
+     */
+    public function beforeTraverse(array $nodes): ?array
+    {
+        $this->classNode = null;
+
+        foreach ($nodes as $node) {
+            if ($node instanceof Class_) {
+                $this->classNode = $node;
+                break;
+            }
+        }
+
+        return null;
+    }
+
+    protected function getClassName(): string
+    {
+        return $this->classNode->namespacedName->toString();
+    }
+
     /**
      * @return null|int|Node
      */
